+++ routes/web.php
@@ -150,7 +150,10 @@ Route::middleware('auth')->group(function () {
         Route::post('/products/{product}/video-tutorials/{tutorial}/toggle', [VideoTutorialController::class, 'toggle'])->name('products.video-tutorials.toggle');
         Route::post('/products/{product}/video-tutorials/reorder', [VideoTutorialController::class, 'reorder'])->name('products.video-tutorials.reorder');
         Route::get('/orders', [OrderController::class, 'index'])->name('orders');
+        Route::get('/orders/members/search', [OrderController::class, 'searchMembers'])->name('orders.members.search');
         Route::post('/orders/{order}/mark-paid', [OrderController::class, 'markPaid'])->name('orders.mark-paid');
+        Route::put('/orders/{order}/affiliate', [OrderController::class, 'updateAffiliate'])->name('orders.update-affiliate');
+        Route::post('/orders/{order}/assign-coupon-owner', [OrderController::class, 'assignCouponOwner'])->name('orders.assign-coupon-owner');
         Route::get('/members', [MemberController::class, 'index'])->name('members');
         Route::get('/members/{user}/edit', [MemberController::class, 'edit'])->name('members.edit');
         Route::put('/members/{user}', [MemberController::class, 'update'])->name('members.update');